pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "fastask"
version = "0.4.7"
description = "The fastest way to get answers in the command line"
authors = [
    {name = "Brian Machado", email = "nah@gmail.com"},
]
readme = "README.md"
dependencies = [
    "requests",
    "openai",
    "python-dotenv",
    "pyyaml"
]

[project.scripts]
ask = "fastask.ask:main"